requirements.txt
aiohttp==3.7.4.post0
aiohttp-retry==2.4.5
appdirs==1.4.4
argon2-cffi==20.1.0
astroid==2.8.0
async-generator==1.10
async-timeout==3.0.1
atpublic==2.0
attrs==20.2.0
autopep8==1.5.7
backcall==0.2.0
bleach==3.2.1
boto3==1.15.17
botocore==1.18.17
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dictdiffer==0.8.1
diskcache==5.2.1
distro==1.5.0
dpath==2.0.5
dulwich==0.20.25
dvc==2.7.4
entrypoints==0.3
flatten-dict==0.4.2
flufl.lock==3.2
fsspec==2021.8.1
ftfy==6.0.3
funcy==1.15
future==0.18.2
gitdb==4.0.5
GitPython==3.1.9
grandalf==0.6
idna==2.10
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.9.3
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
json5==0.9.5
jsonpath-ng==1.5.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab==2.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
kaggle==1.5.12
kiwisolver==1.2.0
lazy-object-proxy==1.6.0
loguru==0.5.3
mailchecker==4.0.12
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
multidict==5.1.0
nanotime==0.5.2
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.1
networkx==2.6.3
notebook==6.1.4
numexpr==2.7.3
numpy==1.21.2
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.1
pathlib2==2.3.5
pathspec==0.8.0
pexpect==4.8.0
phonenumbers==8.12.31
pickleshare==0.7.5
Pillow==8.0.0
platformdirs==2.3.0
ply==3.11
prometheus-client==0.8.0
prompt-toolkit==3.0.8
psutil==5.8.0
ptyprocess==0.6.0
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydot==1.4.1
pygit2==1.6.1
Pygments==2.7.1
pygtrie==2.3.2
pylint==2.11.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-benedict==0.24.2
python-dateutil==2.8.1
python-fsutil==0.5.0
python-slugify==5.0.2
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
requests==2.24.0
rich==10.10.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.7.1
seaborn==0.11.0
Send2Trash==1.5.0
shortuuid==1.0.1
shtab==1.4.1
six==1.15.0
smmap==3.0.4
tables==3.6.1
tabulate==0.8.7
terminado==0.9.1
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
tqdm==4.50.2
traitlets==5.0.5
typing-extensions==3.10.0.2
urllib3==1.25.10
voluptuous==0.12.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.12.1
xmltodict==0.12.0
yarl==1.6.3
zc.lockfile==2.0
zipp==3.3.1
